requirements.txt
pip==19.0.3
python-dateutil==2.8.0
pytz==2019.1
requests==2.21.0
Flask==1.0.2
mysql-connector